An example batch session

Copies the tissue example into a temporary directory n times so the queue has real files to iterate.

Usage

at_example_session(n = 5L, call = rlang::caller_env())

Arguments

n

Number of images in the queue. Default 5.

call

The calling environment, for error reporting.

Value

An annot_session.

See also

Examples

sess <- at_example_session(3)
at_session_status(sess)
#> # A tibble: 3 × 7
#>     idx path                name  status project_path n_rois modified
#>   <int> <chr>               <chr> <chr>  <chr>         <int> <dttm>  
#> 1     1 /tmp/RtmpfWJMsT/an… imag… pendi… NA                0 NA      
#> 2     2 /tmp/RtmpfWJMsT/an… imag… pendi… NA                0 NA      
#> 3     3 /tmp/RtmpfWJMsT/an… imag… pendi… NA                0 NA